% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{color_maps}
\alias{color_maps}
\title{Color Maps for Plotting}
\format{
An object of class \code{list} of length 3.
}
\usage{
color_maps
}
\description{
A list of color maps that can be used for plotting
}
\keyword{datasets}
