% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{color_palettes}
\alias{color_palettes}
\title{Color Palettes for Plotting}
\format{
An object of class \code{list} of length 2.
}
\usage{
color_palettes
}
\description{
Color Palettes for Plotting
}
\keyword{datasets}
