% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnplot_reference.R
\name{columnplot_reference}
\alias{columnplot_reference}
\title{Title}
\usage{
columnplot_reference(
  data,
  projections = paste0("Q", 1:7),
  SuperPopulation = SuperPopulation,
  Population = Population,
  population_label = "1KG Population"
)
}
\arguments{
\item{population_label}{}
}
\description{
Title
}
