% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{create_tooltips}
\alias{create_tooltips}
\title{Title}
\usage{
create_tooltips(
  data,
  header = NULL,
  include = NULL,
  digits = getOption("digits")
)
}
\arguments{
\item{data}{A data frame to evaluate tooltips in the context of.}

\item{header}{A glue string that will form the tooltip header.}

\item{include}{A list of column names to include in the tooltip. tidyselect syntax is
supported.}
}
\description{
TODO: Incorporate labelled data for the names, instead of variable name
}
