% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecla.R
\docType{data}
\name{ecla}
\alias{ecla}
\title{Estimated Cell Line Ancestry Data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3897 rows and 45 columns.
}
\usage{
ecla
}
\description{
Data used for the Cancer Research paper on estimated cell line
ancestry.
}
\keyword{datasets}
