% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{map_1kg_population}
\alias{map_1kg_population}
\title{Map for 1KG Population Labels}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 26 rows and 9 columns.
}
\usage{
map_1kg_population
}
\description{
Map for 1KG Population Labels
}
\keyword{datasets}
