% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{map_1kg_superpopulation}
\alias{map_1kg_superpopulation}
\title{Map for 1KG SuperPopulation Labels}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 5 rows and 2 columns.
}
\usage{
map_1kg_superpopulation
}
\description{
Map for 1KG SuperPopulation Labels
}
\keyword{datasets}
