% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{plot_interactive}
\alias{plot_interactive}
\title{Add interactivity to ggplot}
\usage{
plot_interactive(x)
}
\arguments{
\item{x}{A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.}
}
\value{
A \code{\link[ggiraph:girafe]{ggiraph::girafe()}} object for visualizing.
}
\description{
This function will wrap the ggplot provided into an interactivity plot
via the \code{\link[ggiraph:girafe]{ggiraph::girafe()}} function. This function will setup the
options specific for this package.
}
