% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pull}
\alias{pull}
\title{Pull column corresponding to variable name}
\usage{
pull(.x, v)
}
\arguments{
\item{v}{}
}
\description{
Similar to \code{\link[dplyr:pull]{dplyr::pull()}}, but for formula-based
naming of variables (for plotly).
}
