% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot_overlay.R
\name{scatterplot_overlay}
\alias{scatterplot_overlay}
\title{Title}
\usage{
scatterplot_overlay(
  p,
  data,
  x = NULL,
  y = NULL,
  label = NULL,
  subset = NULL,
  color_by = NULL,
  tooltip_header = NULL,
  tooltip_include = NULL,
  tooltip_digits = getOption("digits"),
  alpha = 0.2
)
}
\arguments{
\item{alpha}{}
}
\description{
Title
}
