% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot_reference.R
\name{scatterplot_reference}
\alias{scatterplot_reference}
\title{Generate Scatter Plot of Reference Population}
\usage{
scatterplot_reference(
  p = NULL,
  data,
  x = NULL,
  y = NULL,
  label = NULL,
  subset = NULL,
  tooltip_header = NULL,
  tooltip_include = NULL,
  tooltip_digits = getOption("digits"),
  alpha = 0.2,
  color_by = NULL,
  color_map = NULL
)
}
\arguments{
\item{p}{A ggplot object (possibly NULL) to add reference samples. If NULL, a new
plot will be created.}

\item{data}{A data frame to extract reference information from for plotting. Many
of the subsequent parameters are columns from this data to use for various purposes.}

\item{x}{The x coordinate column of the projection}

\item{y}{The y coordinate column of the projection}

\item{label}{A column to use as labels for the reference population. Note that a
reference population is typically clustered. This column should have repeated
labels for every member of the cluster. A representative sample will be identified
among rows having the same label value (currently random) and labelled on the graph
using \code{\link[ggrepel:geom_text_repel]{ggrepel::geom_text_repel()}}.

If this value is NULL (default) when using the \code{color_by} parameter, then the
\code{color_by} values will be printed as labels. This is not desirable in the case
of a continuous value \code{color_by} option.}

\item{subset}{Subset of samples to plot using color scheme (all others will be gray)}

\item{tooltip_header}{A \code{\link[glue:glue]{glue::glue()}} string to use as the header within a tooltip. Most
typically used to refer to specific columns in \code{data} when creating the tooltip header.}

\item{tooltip_include}{A list of columns in \code{data} to include within the tooltip text.}

\item{tooltip_digits}{The number of digits when formatting numbers in the tooltips. Default
is \code{geOption("digits")}.}

\item{alpha}{(default 0.2) The alpha level for reference samples. Too high a value and it
becomes difficult to see the overlay samples.}

\item{color_by}{A column to use for varying colors among the populations. Typically this
would be a Population or SuperPopulation option, however continuous variables can also
be used.}

\item{color_map}{A named vector of colors to use when assigning colors to data points.
This is tied to the \code{color_by} parameter and will assign colors to specific values pointed
to by that parameter. When continuous, this parameter is used for the max value (the min
value is set to gray currently).}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object representing the reference cohort projected onto
the \code{x} and \code{y} values.
}
\description{
This function will plot data from a data frame, under the assumption that the
data frame represents various pieces of information relevant for estimated
ancestry (or human diversity map) within a reference population.
}
\details{
A variety of terms are important here:
\itemize{
\item Reference Population: There are several efforts to measure human variation
by evaluating specific populations across the globe. Examples include the
Thousand Genomes Project (1KG) or HGDP (Human Genome Diversity Project).
\item Projection: Genetic diversity is a genome (typically exome)-wide evaluation
of human variation. However, to visualize this variability we need to project
this variation into a relatively small number of coordinates (typically 2) that
can be plotted.
}
}
\note{
Most parameters default to NULL so that intelligent (ggplot) defaults can
be used unless overridden.
}
\examples{
\dontrun{
scatterplot_reference(
  data = iris,
  x = Sepal.Length,
  y = Sepal.Width,
  color_by = Species,
  color_map = c("red","green","blue")
)
}
}
